Created script to copy keymap to config subdirectories
The ZMK Keymap Editor (https://nickcoutsos.github.io/keymap-editor/)
requires that the `.keymap` file be in the `config` directory.

The working copy that the configurator uses will be in the `config`
directory, and immediately after committing the changes made with the
configurator, run this script so the keymap ends up in the correct
subdirectories.

You could also copy the keymap file yourself, but this makes the process
just a little bit easier.

Ideally, we can add something to CI such that after a change is made to
the `config/cyber_60_rev_c.keymap` file, it will run `copy-layout.sh`
before running the job -- but we can handle that later.